requirements.txt
Django>=3.2
djangorestframework>=3.12
psycopg2-binary>=2.8
pytest>=6.0
pytest-django>=4.0